Refactor so that only subsets of tests get run
