--- .github/workflows/ci.yaml
@@ -1,42 +1,0 @@
-name: "CI"
-on: [pull_request, push]
-
-jobs:
-  lint:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v2
-      - uses: actions/setup-python@v2
-        with:
-          python-version: "3.8"
-      - uses: pre-commit/action@v2.0.0
-
-  test:
-    runs-on: ubuntu-latest
-    strategy:
-      fail-fast: false
-      matrix:
-        python-version: ["3.8", "3.9", "3.10"]
-        suite: ["", "helm", "operator"]
-    steps:
-      - uses: actions/checkout@v2
-      - uses: actions/setup-python@v2
-        with:
-          python-version: ${{ matrix.python-version }}
-
-      - name: Install deps
-        run: |
-          pip install -e .
-          pip install -r requirements-test.txt
-          pip install git+https://github.com/dask/distributed@main
-          pip install git+https://github.com/dask/dask@main
-
-      - name: Prebuild long running session fixtures
-        env:
-          KUBECONFIG: .pytest-kind/pytest-kind/kubeconfig
-        run: pytest dask_kubernetes/tests/test_async.py::test_fixtures
-
-      - name: Run tests
-        env:
-          KUBECONFIG: .pytest-kind/pytest-kind/kubeconfig
-        run: pytest --durations=10 dask_kubernetes/${{ matrix.suite }}/tests
