+++ .github/workflows/helmcluster-ci.yaml
@@ -1,0 +1,30 @@
+name: "HelmCluster CI"
+on:
+  pull_request:
+    paths: &helmcluster_paths
+      - "ci/**"
+      - "dask_kubernetes/helm/**"
+      - "dask_kubernetes/*"
+  push:
+    paths: *helmcluster_paths
+
+
+jobs:
+  test:
+    runs-on: ubuntu-latest
+    strategy:
+      fail-fast: false
+      matrix:
+        python-version: ["3.8", "3.9", "3.10"]
+
+    steps:
+      - uses: actions/checkout@v2
+      - uses: actions/setup-python@v2
+        with:
+          python-version: ${{ matrix.python-version }}
+      - name: Install deps
+        run: ./ci/install-deps.sh
+      - name: Run tests
+        env:
+          KUBECONFIG: .pytest-kind/pytest-kind/kubeconfig
+        run: pytest dask_kubernetes/helm/tests
