+++ dask_kubernetes/classic/__init__.py
@@ -1,0 +1,1 @@
+from .kubecluster import KubeCluster
