+++ setup.cfg
@@ -36,7 +36,7 @@ tag_prefix =
 parentdir_prefix = dask-kubernetes-
 
 [tool:pytest]
-addopts = -v --keep-cluster
+addopts = -v --keep-cluster --durations=10
 timeout = 300
 reruns = 5
 asyncio_mode = auto
